Move random scripts to tools
